WIRE: Correct (increase) delay in write/read slot, DELAY_SLOTE
Increase delay in write/read slot to work protocol correctly. After
changing compiler from IAR to gcc the temperature sensor, i.e
DS18B20, doesn't work. The protocol specification claims to take delay
about 60us, but we set about 18us. So to adjust delay close to
claimed set it about 40us wich is more actually because of delay
between function calls. If it will be needed more it should be
increased, but at that moment it is taken with spare.

Signed-off-by: Ivan Khoronzhuk <khorivan@gmail.com>