BC1602E: mark internal functions as static
In IAR marking internal functions as static was not needed.
But now it is correct to use static when options -ffunction-sections
and -fdata-sections are not set. For clarity reasons and to make code
similar like in LK "static" option is added when it is can be needed.

Also remove redundant comments.

Signed-off-by: Ivan Khoronzhuk <khorivan@gmail.com>